addressed the PR comments
